+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Added glossary.
 - Added theme props, with support for "light" and "dark". Added a URL theme parameter for the demo.
 
+### Changed
+- Upgrade vitessce-image-viewer to 0.1.3 & use data loaders
+
 ## [0.0.24](https://www.npmjs.com/package/vitessce/v/0.0.24) - 2020-03-02
 ### Added
 - Check that HTTP status is good before trying to parse response.