+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Upgrade vitessce-image-viewer to 0.1.3 & use data loaders.
 - Rename "LayerPublisher" to "SourcePublisher".
 - Converted the polygon tool to a lasso tool.
+- Loosened assumptions about cell info object properties when rendering cell hover messages to fix a `TypeError` bug.
 
 ## [0.0.24](https://www.npmjs.com/package/vitessce/v/0.0.24) - 2020-03-02
 ### Added